Fix compiler warning in the test suite from amd64 system
size_t is not an int.

Reviewed-on: https://gerrit.stanford.edu/134
Tested-by: Russ Allbery <rra@stanford.edu>

+++ tests/lib/attrs-t.c
@@ -88,7 +88,7 @@ main(void)
            "...with the right count");
     for (i = 0; i < attrs_out->num_attrs; i++) {
         is_string(attrs_in->attrs[i].name, attrs_out->attrs[i].name,
-                  "Attribute %d has the right value", i);
+                  "Attribute %lu has the right value", (unsigned long) i);
         is_int(attrs_in->attrs[i].length, attrs_out->attrs[i].length,
                "...and the right length");
         ok(memcmp(attrs_in->attrs[i].value, attrs_out->attrs[i].value,
